requirements.txt
canvasapi==3.2.0
discord.py==2.3.2
html2text==2020.1.16
ics==0.7.2
python-dotenv==1.0.0
